requirements.txt
numpy
pandas
pdfplumber
pytesseract
sentence-transformers
faiss-cpu
langchain_community
langgraph
typing
langchain-openai
langchain-core
langchain
langchain-text-splitters
notebook
python-box
fastapi
uvicorn
Jinja2
